requirements.txt
streamlit==1.25.0
snowflake-sqlalchemy
snowflake-connector-python[pandas]